backend/requirements.txt
fastapi
uvicorn
pydriller
GitPython
pandas